+++ server/src/compiler_analyzer/symbolComplement.ts
@@ -2,7 +2,7 @@ import {ParsedRange} from "../compiler_parser/nodes";
 import {getNodeLocation} from "../compiler_parser/nodesUtils";
 import {LocationInfo} from "../compiler_tokenizer/tokens";
 import {ParserToken} from "../compiler_parser/parserToken";
-import {SymbolType, SymbolFunction} from "./symbols";
+import {SymbolType, SymbolFunction} from "./symbolObject";
 import {TemplateTranslation} from "./symbolUtils";
 import {SymbolScope} from "./symbolScope";
