+++ server/src/compiler_analyzer/symbolObject.ts
